Refactored ActivitiesLogController, enhanced financial report handling, improved product controllers, updated user model and authentication logic, and added policy for financial reports

+++ app/Http/Controllers/ProductCategoryController.php
@@ -78,6 +78,8 @@ class ProductCategoryController extends Controller
      */
     public function destroy(ProductCategory $productCategory): JsonResponse
     {
+        $this->authorize('deleteProductCategory', ProductCategory::class);
+
         $result = $this->productcategoryService->deleteProductCategory($productCategory);
 
         return $result['status'] === 200
